add checkpoint size utilities

+++ src/dftracer/utils/indexer/helpers.h
@@ -5,13 +5,9 @@
 #include <string>
 
 std::string get_logical_path(const std::string &path);
-
 time_t get_file_modification_time(const std::string &file_path);
-
 std::string calculate_file_sha256(const std::string &file_path);
-
-std::size_t file_size_bytes(const std::string &path);
-
+std::uint64_t file_size_bytes(const std::string &path);
 bool index_exists_and_valid(const std::string &idx_path);
 
 #endif  // DFTRACER_UTILS_INDEXER_HELPERS_H
